landing page, near, and played working
In this, I turned the app.jsx file into a landing page and added conditions for the text to appear. I then added two new pages "near" and "played". These represent the two pages that will be primarily used throughout the project. I need to probably do something with firebase and gps tracking but that'll come in the future